+++ example/index.js
@@ -44,7 +44,7 @@ console.log(temp.trimRight(), "\n");
 console.log("Finding an object with property 'country' as 'NI'");
 console.log(db.find((u) => u.country === "NI"), "\n");
 
-console.log("Filtering objects with property 'name'");
+console.log("Filtering objects with property 'name' (all objects have it*)");
 console.log(db.filter((u) => u.name), "\n");
 
 console.log("Deleting an object");