+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 ## Not Published
 - Added `find` and `filter` functions
 - `README.md` more beautiful
-- ~~`create` function now adds nonexisting properties~~
+- The ID of an element now supports breaklines (\n)
 
 ## 2.1.4 - Error fixed
 - `push` in an array (MeowDBObject property) works now.